Added message analyzer for line, to check for absence messages
